% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black_box_explain.R
\name{black_box_explain}
\alias{black_box_explain}
\title{Black Box explain}
\usage{
black_box_explain()
}
\description{
Shiny app permettant d'analyser l'interprétabilité et les prédictions obtenues
d'un modèle (en important le modèle mlr et le jeu de données dans l'application)
}
\author{
Gabriel Crépeault-Cauchon
}
