% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsPredDistribution.R
\name{bsPredDistribution}
\alias{bsPredDistribution}
\title{bsPredDistribution}
\usage{
bsPredDistribution(learner, train_tsk, test_tsk, B = 100)
}
\arguments{
\item{learner}{Leaner mlr utilisé pour fiter les modèles}

\item{train_tsk}{Tâche mlr de validation}

\item{B}{Nombre d'échantillonage Bootstrap à effectuer. Défaut = 100}
}
\value{
Une liste contenant :
- predictions : un data.frame contenant une ligne par
observations, une colonne par modèle Bootstrap.
- bs_models : liste contenant tous les modèles (provenant du resampling) Bootstrap.
}
\description{
Fonction wrapper qui effectue un rééchantillonage Bootstrap sur le
training set et prédit le validation set B fois. De cette façon, on obtient B estimations
pour chaque observation dans le validation set.
}
\author{
Gabriel Crépeault-Cauchon
}
