% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censoreDataset.R
\name{censoreDataset}
\alias{censoreDataset}
\title{censoreDataset}
\usage{
censoreDataset(data, evalYr, rep.value = NA)
}
\arguments{
\item{data}{Base de données utilisées. Doit avoir un certain format}

\item{evalYr}{Date d'évaluation}

\item{rep.value}{par quelle valeur on remplace les valeurs qu'on ne connaît pas?}

\item{keep.real.ultimate}{Indicateur pour dire si on veut ajouter une colonne real_ultimate, qui contiendra
la somme de tous les paiements faits (sur toutes les périodes de développement) pour chaque réclamation.}
}
\value{
Un datasets qui rend non-disponible certaines informations (qui sont normalement
non dispo à la date d'évaluation).
}
\description{
Une fonction qui vient réduire l'information qu'on est censé avoir
de disponible à une certaine date d'évaluation
}
\author{
Gabriel Crépeault-Cauchon
}
