% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAggrTriangle.R
\name{getAggrTriangle}
\alias{getAggrTriangle}
\title{getAggrTriangle}
\usage{
getAggrTriangle(data, pmtVar = "Pay")
}
\arguments{
\item{data}{jeu de données provenant des données simulées. Au minimum, le jeu de données doit avoir
une colonne AY (Accident Year) et des colonnes de paiement incrémentaux (avec un préfixe commun, e.g. Pay<XX>)}

\item{pmtVar}{préfixe dans le nom des variables du dataset qui contiennent les paiements.}
}
\description{
Fonction qui permet de prendre un datasets provenant de la machine à simulation de
<mettre le nom de l'auteur> et d'en obtenir un triangle de paiements cumulatifs (utilisé
par la suite dans des techniques ChainLadder).
}
