% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLDF.R
\name{getLDF}
\alias{getLDF}
\title{getLDF}
\usage{
getLDF(tri, tailLDF = 1, output = "ldf", dec = 4)
}
\arguments{
\item{tri}{Un object triangle du package ChainLadder}

\item{tailLDF}{hypothèse prise pour le dernier LDF calculé (tail), soit le
développement final}

\item{output}{Chaine de caractères pour préciser le type d'output désiré (seulement
les age-to-age ratios 'ata', seulement les LDF 'ldf' ou une liste contenant les deux 'both')}
}
\description{
Fonction qui permet d'obtenir les age-to-age developpement factor (ldf) à partir
d'un triangle aggrégé
}
