% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_selon_variable.R
\name{rmse_selon_variable}
\alias{rmse_selon_variable}
\title{rmse_selon_variable}
\usage{
rmse_selon_variable(pred, var, show.table = F)
}
\arguments{
\item{pred}{Un data.frame contenant, au minimum :
- "truth" : la vraie valeur à prédire
- "response" : valeur prédite
- "var" : la variable explicative qu'on veut illustrer sur le graphique.}

\item{var}{Nom de la variable (avec syntaxe tidyverse) qu'on veut grouper}

\item{show.table}{Possbilité de print le data.frame aggrégé (qui présente le nombre
d'observation et le RMSE moyen) qui a servi à produire le graphique.}
}
\description{
Fonction wrapper pour obtenir facilement le graphique du RMSE moyen en
groupant sur l'une des variables explicatives. Le graphique contient aussi un histogramme
mentionnant le nombre d'observations par catégorie de la variable explicative.
}
