% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_training_env.R
\name{setup_training_env}
\alias{setup_training_env}
\title{setup_training_env}
\usage{
setup_training_env(model_name, root = file.path("models", model_name))
}
\arguments{
\item{model_name}{}
}
\description{
Monter un dossier avec des sous-dossier pour organiser
l'entrainement d'un modèle
}
