% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimulatedIndClaims}
\alias{SimulatedIndClaims}
\title{Simulated data from the}
\format{R object of class `data.frame` with 37190 rows and 32 columns :
\describe{
\item{ClNr}{Artificial Clainumber, 1 to n}
\item{LoB}{Line of business}
\item{cc}{Claims code}
\item{AY}{Accident Year}
\item{AQ}{Accident Quarter}
\item{age}{Age of the insured, categorical}
\item{inj_part}{Injured part, categorical}
\item{RepDel}{Reporting Delay}
\item{Pay<xx>}{Incremental Payment. A column for each developpement year i, i = 0, ..., 11}
\item{Open<xx>}{Status of claim (1 = Open, 0 = Closed). A column for each developpement year i, i = 0, ..., 11}
}}
\source{
Machine à simulation de Gabrielli/Wuthrich:2018
}
\usage{
SimulatedIndClaims
}
\description{
Jeu de données simulées de réclamations individuelles d'une compagnie d'assurance
de 1994 à 2005.
}
\keyword{datasets}
