% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_and_update.R
\name{tune_and_update}
\alias{tune_and_update}
\title{tune_and_update}
\usage{
tune_and_update(learner, task, param_set, plot = T, show.info = T, root)
}
\arguments{
\item{learner}{Un objet R de classe "Learner"}

\item{task}{Un objet R de class "Task"}

\item{param_set}{Un objet R de classe "ParamSet"}

\item{plot}{indicateur pour indiquer si on veut produire un graphique résumé 
de la procédure CV5}

\item{export}{indicateur pour indiquer si on veut sauvegarder dans un fichier .rda
l'objet TuneResult (lorsque c'est long à rouler, c'est intéressant de sauvegarder des
résultats intermédiaires.}
}
\value{
Une liste contenant un graphique, l'objet TuneResult, le data.frame ayant
permi de créer le graphique et le nouvel objet Learner mis à jour avec la valeur optimale
du paramètre.
}
\description{
Fonction pour wrapper la fonction tuneParams de mlr avec les différents
paramètres à fixer. Ça réduit le code dans le fichier principal de tuning.
}
\author{
Gabriel Crépeault-Cauchon
}
