% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_nrounds.R
\name{tune_nrounds}
\alias{tune_nrounds}
\title{tune_nrounds}
\usage{
tune_nrounds(
  learner,
  task,
  root,
  iters = 1000,
  plot = T,
  export = T,
  early_stopping_rounds = 10
)
}
\description{
Fonction wrapper pour utilier la fonction xgb.cv du package xgboost
}
\details{
TODO : description de la fonction et ses paramètres
}
