% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widetolong.R
\name{widetolong}
\alias{widetolong}
\title{widetolong}
\usage{
widetolong(dt, ldf, pmtVar = "Pay", statusVar = "Open")
}
\arguments{
\item{dt}{Jeu de données}

\item{pmtVar}{Préfixe des variables de paiement}

\item{statusVar}{Préfixe des variables qui indique le statut de la réclamation}

\item{dup.rm}{boolean : TRUE si on veut retirer les lignes duplicate}
}
\description{
Fonction qui permet d'éclater un 
dataset pour avoir l'évolution de la réclamation à chaque année 
de développement. La fonction nous retourne les lignes sans doublons
}
\author{
Gabriel Crépeault-Cauchon
}
